requirements.txt
pandas>=1.5.0
yfinance>=0.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.3.0
